refresh detector metadata on model update
