+++ app/core/app_state.py
@@ -1,24 +1,15 @@
 import logging
 import os
-from functools import lru_cache
 
-import cachetools
 import yaml
 from fastapi import Request
-from groundlight import Groundlight
-from model import Detector
 
 from .configs import EdgeInferenceConfig, RootEdgeConfig
 from .database import DatabaseManager
 from .edge_inference import EdgeInferenceManager
 from .file_paths import DEFAULT_EDGE_CONFIG_PATH
-from .utils import safe_call_sdk
 
 logger = logging.getLogger(__name__)
-
-MAX_SDK_INSTANCES_CACHE_SIZE = 1000
-MAX_DETECTOR_IDS_TTL_CACHE_SIZE = 1000
-TTL_TIME = 3600  # 1 hour
 
 
 def load_edge_config() -> RootEdgeConfig:
@@ -79,34 +70,6 @@ def get_detector_inference_configs(
         }
 
     return detector_to_inference_config
-
-
-@lru_cache(maxsize=MAX_SDK_INSTANCES_CACHE_SIZE)
-def _get_groundlight_sdk_instance_internal(api_token: str):
-    return Groundlight(api_token=api_token)
-
-
-def get_groundlight_sdk_instance(request: Request):
-    """
-    Returns a (cached) Groundlight SDK instance given an API token.
-    The SDK handles validation of the API token token itself, so there's no
-    need to do that here.
-    """
-    api_token = request.headers.get("x-api-token")
-    return _get_groundlight_sdk_instance_internal(api_token)
-
-
-@cachetools.cached(
-    cache=cachetools.TTLCache(maxsize=MAX_DETECTOR_IDS_TTL_CACHE_SIZE, ttl=TTL_TIME),
-    key=lambda detector_id, gl: detector_id,
-)
-def get_detector_metadata(detector_id: str, gl: Groundlight) -> Detector:
-    """
-    Returns detector metadata from the Groundlight API.
-    Caches the result so that we don't have to make an expensive API call every time.
-    """
-    detector = safe_call_sdk(gl.get_detector, id=detector_id)
-    return detector
 
 
 class AppState:
